[MySQL] Ee-enable ALPHANUMERIC string generation for PQS by splitting up the test classes
