+++ .travis.yml
@@ -59,7 +59,8 @@ matrix:
         - sudo apt-get update && sudo apt-get install libssl-dev libmecab2 libjson-perl mecab-ipadic-utf8
         - sudo apt-get remove mysql-* && wget https://dev.mysql.com/get/Downloads/MySQL-8.0/mysql-server_8.0.20-1ubuntu18.04_amd64.deb-bundle.tar && tar -xvf mysql-server_8.0.20-1ubuntu18.04_amd64.deb-bundle.tar && yes | sudo dpkg -i *.deb
         - sudo mysql -e "CREATE USER 'sqlancer'@'localhost' IDENTIFIED BY 'sqlancer'; GRANT ALL PRIVILEGES ON * . * TO 'sqlancer'@'localhost';"
-        - MYSQL_AVAILABLE=true mvn -Dtest=TestMySQL test
+        - MYSQL_AVAILABLE=true mvn test -Dtest=TestMySQLPQS
+        - MYSQL_AVAILABLE=true mvn test -Dtest=TestMySQLTLP
     - name: CockroachDB
       jdk : openjdk8
       before_install:
